Add files via upload
A big table was created, from which the ERD table for expert reviews was created (tiny table). The datep was changed to the date datetype, and the idvscore was changed to numeric (5,2). Idvscore was made not null, with two absent values filled in as 0. PK and FK were not made as another table is needed for the connection, and [url, reviewer, datep, rev] cannot be PK as they are not unique.